requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
google-generativeai==0.3.0
python-dotenv==1.0.0
email-validator==2.0.0
python-dotenv==1.0.0